pyproject.toml
[project]
name = "youtube-rag-groq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "ffmpeg>=1.4",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langchain-pinecone==0.2.13",
    "openai>=2.24.0",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pinecone>=7.3.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "yt-dlp>=2026.3.3",
]
